backend/requirements.txt
# FastAPI and related
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
pydantic[email]==1.8.2

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0

# Supabase
gotrue==0.2.0
postgrest==0.4.0
realtime==0.1.0
supabase==0.7.1

# Environment variables
python-dotenv==0.19.0

# CORS
starlette==0.14.2

# Utilities
requests==2.26.0
uuid==1.30
